Example app: don't keep fetching room
Split all the room stuff out into a RoomView and fetch the room only
once.
